// src/app/models/votingDetails.ts
export interface votingDetails{
    BestSpeaker: string;
    BestEvaluator: string;
    BestTTSpeaker: string;
    BestTAGPlayer: string;
    BestRolePlayer: string;
}

export interface ResultsDetails{
    voteCount: number;
    id: string;
}
export interface votingResultsURL {
    BestSpeakerURL: string;
    BestEvaluatorURL: string;
    BestTTSpeakerURL: string;
    BestTAGPlayerURL: string;
    BestRolePlayerURL: string;
}